src/app/components/top-bar/top-bar.component.ts
import { Component, EventEmitter, Output } from '@angular/core';
import { AuthService } from '../../auth/auth.service';
import { Router, RouterLink } from '@angular/router';
import { MatToolbarModule } from '@angular/material/toolbar';
import { MatIconModule } from '@angular/material/icon';
import { MatButtonModule } from '@angular/material/button';
import { MatSidenavModule } from '@angular/material/sidenav';

@Component({
  selector: 'app-top-bar',
  standalone: true,
  imports: [
    RouterLink,
    MatToolbarModule,
    MatIconModule,
    MatButtonModule,
    MatSidenavModule,
  ],
  templateUrl: './top-bar.component.html',
  styleUrl: './top-bar.component.css'
})
export class TopBarComponent {

  @Output() menuButtonPressed = new EventEmitter<boolean>();
  menuIsOpen = false;

  constructor(private auth: AuthService, private router: Router) { }

  menuButtonClick(): void { 
    this.menuButtonPressed.emit(this.menuIsOpen = !this.menuIsOpen); 
  }
}
